+++ private/build_mutation_fields/index.js
@@ -5,11 +5,15 @@ const {
   GraphQLInputObjectType,
   GraphQLList
 } = require('graphql')
+const broadcast_resolver = require('../resolvers/broadcast.js')
+const resolver = require('../resolvers/index.js')
 const serialize_transaction_data = require('../serialize/transaction_data.js')
 const ast_to_input_types = require('./ast_to_input_types.js')
 const authorization_type = require('./types/authorization_type.js')
 const configuration_default_value = require('./types/configuration_default_value.js')
 const configuration_type = require('./types/configuration_type.js')
+const packed_transaction_type = require('./types/packed_transaction_type.js')
+const transaction_receipt_type = require('./types/transaction_receipt_type.js')
 
 /**
  * Builds a GraphQL mutation input field from an `abi ast`.
@@ -17,13 +21,12 @@ const configuration_type = require('./types/configuration_type.js')
  * signed and broadcast to the EOSIO blockchain or you can just return the serialization type.
  * @name build_mutation_fields
  * @kind function
- * @param {object} abi_ast Abstract syntax tree (AST) for a given smart contract.
- * @param {Function} resolver A resolver for handling EOSIO transaction.
- * @param {object} type A GraphQL return type.
+ * @param {ABI_AST} abi_ast Abstract syntax tree (AST) for a given smart contract.
+ * @param {bool} broadcast Determines if transaction is pushed to the blokchain or just serialized.
  * @returns {object} GraphQL mutation fields.
  * @ignore
  */
-function build_mutation_fields(abi_ast, resolver, type) {
+function build_mutation_fields(abi_ast, broadcast) {
   const ast_input_object_types = ast_to_input_types(abi_ast)
 
   const fields = abi_ast.actions.reduce(
@@ -51,12 +54,12 @@ function build_mutation_fields(abi_ast, resolver, type) {
 
   return {
     [abi_ast.gql_contract]: {
-      type,
+      type: broadcast ? transaction_receipt_type : packed_transaction_type,
       description: `Update the \`${abi_ast.contract}\` smart contract.`,
       args: {
         actions: {
           type: new GraphQLList(
-            GraphQLNonNull(
+            new GraphQLNonNull(
               new GraphQLInputObjectType({
                 name: `${abi_ast.gql_contract}_action`,
                 description: `List of mutations on the ${abi_ast.contract} smart contract.`,
@@ -109,7 +112,7 @@ function build_mutation_fields(abi_ast, resolver, type) {
           }
         }
 
-        return resolver(
+        const packed_transaction = await resolver(
           {
             configuration,
             actions: _actions,
@@ -119,6 +122,14 @@ function build_mutation_fields(abi_ast, resolver, type) {
           },
           abi_ast
         )
+
+        if (broadcast)
+          return broadcast_resolver({
+            ...packed_transaction,
+            private_keys,
+            rpc_url
+          })
+        else return packed_transaction
       }
     }
   }