+++ private/build_mutation_fields/types/authorization_type.js
@@ -3,6 +3,15 @@
 const { GraphQLInputObjectType, GraphQLNonNull } = require('graphql')
 const name_type = require('../../eosio_types/name_type')
 
+/**
+ * The action authorization type for action validation.
+ * @kind typedef
+ * @name authorization
+ * @type {object}
+ * @prop {string} actor Name of the account that is trying to authorize.
+ * @prop {string} permission Name of the `permission` of the the `actor`
+ */
+
 const authorization_type = new GraphQLNonNull(
   new GraphQLInputObjectType({
     name: 'authorization',