+++ private/build_mutation_fields/types/configuration_type.js
@@ -2,6 +2,19 @@
 const { GraphQLInputObjectType, GraphQLInt } = require('graphql')
 const configuration_default_value = require('./configuration_default_value.js')
 
+/**
+ * Handles various elements relating to the `EOSIO` transaction.
+ * @kind typedef
+ * @name configuration_type
+ * @type {object}
+ * @prop {number} blocksBehind Number of blocks behind the current block (TaPoS protection).
+ * @prop {number} expireSeconds  Seconds past before transaction is no longer valid (TaPoS protection)
+ * @prop {number} max_net_usage_words  Maximum NET bandwidth usage a transaction can consume (0 there is no limit).
+ * @prop {number} max_cpu_usage_ms  Maximum NET bandwidth usage a transaction can consume (0 there is no limit).
+ * @prop {number} delay_sec  Number of seconds that the transaciton will be delayed by.
+ * @ignore
+ */
+
 const configuration_type = new GraphQLInputObjectType({
   name: 'configuration_type',
   description: