+++ private/eosio_types/block_timestamp_type.js
@@ -8,7 +8,7 @@ const block_timestamp_type = new GraphQLScalarType({
   parseValue: block_time_stamp => {
     if (block_time_stamp == '') return ''
     if (Number.isNaN(Date.parse(block_time_stamp)))
-      throw new Error('Invalid block timestamp')
+      throw new TypeError('Invalid block timestamp ' + block_time_stamp)
     return block_time_stamp
   }
 })