+++ private/eosio_types/signature_type.js
@@ -9,10 +9,10 @@ const signature_type = new GraphQLScalarType({
   parseValue: signature => {
     if (signature == '') return ''
     if (typeof signature !== 'string')
-      throw new Error('Expected signature to be string')
+      throw new TypeError('Expected signature to be string')
     const prefix = signature.substr(0, 7)
     if (prefix != 'SIG_K1_')
-      throw new Error('Signature prefix is should be SIG_K1_.')
+      throw new TypeError('Signature prefix is should be SIG_K1_.')
     return signature
   }
 })