+++ test/server.test.js
@@ -1,7 +1,7 @@
 'use strict'
 
 const http = require('http')
-const { SmartQL } = require('../public/index.js')
+const SmartQL = require('../private/index.js')
 
 const server = http
   .createServer(function (request, response) {
@@ -11,10 +11,9 @@ const server = http
 
       SmartQL({
         query,
-        broadcast: false,
+        broadcast: true,
         rpc_url: 'https://jungle.relocke.io',
-        contract: 'eosio',
-        private_keys: ['5K7xR2C8mBzMo4aMPJyBPp7Njc3XvszeJSfTApa51rc2d54rrd3']
+        contracts: ['eosio', 'eosio.token', 'relocke']
       }).then(async data => {
         response.writeHead(200, { 'Content-Type': 'application/json' })
         response.end(JSON.stringify(data), 'utf-8')